1.1.0 update

+++ src/main/java/net/lomeli/lomlib/client/render/FluidRender.java
@@ -26,26 +26,25 @@ public class FluidRender {
     private static final BlockInterface liquidBlock = new BlockInterface();
 
     public static Icon getFluidTexture(FluidStack fluidStack, boolean flowing) {
-        if(fluidStack == null) {
+        if (fluidStack == null) {
             return null;
         }
         return getFluidTexture(fluidStack.getFluid(), flowing);
     }
 
     public static Icon getFluidTexture(Fluid fluid, boolean flowing) {
-        if(fluid == null) {
+        if (fluid == null) {
             return null;
         }
         Icon icon = flowing ? fluid.getFlowingIcon() : fluid.getStillIcon();
-        if(icon == null) {
-            icon = ((TextureMap) Minecraft.getMinecraft().getTextureManager().getTexture(TextureMap.locationBlocksTexture))
-                    .getAtlasSprite("missingno");
+        if (icon == null) {
+            icon = ((TextureMap) Minecraft.getMinecraft().getTextureManager().getTexture(TextureMap.locationBlocksTexture)).getAtlasSprite("missingno");
         }
         return icon;
     }
 
     public static ResourceLocation getFluidSheet(FluidStack liquid) {
-        if(liquid == null)
+        if (liquid == null)
             return BLOCK_TEXTURE;
         return getFluidSheet(liquid.getFluid());
     }
@@ -55,36 +54,36 @@ public class FluidRender {
     }
 
     public static void setColorForFluidStack(FluidStack fluidstack) {
-        if(fluidstack == null)
+        if (fluidstack == null)
             return;
 
         int color = fluidstack.getFluid().getColor(fluidstack);
-        float red = (float) (color >> 16 & 255) / 255.0F;
-        float green = (float) (color >> 8 & 255) / 255.0F;
-        float blue = (float) (color & 255) / 255.0F;
+        float red = (color >> 16 & 255) / 255.0F;
+        float green = (color >> 8 & 255) / 255.0F;
+        float blue = (color & 255) / 255.0F;
         GL11.glColor4f(red, green, blue, 1);
     }
 
     public static int[] getFluidDisplayLists(FluidStack fluidStack, World world, boolean flowing) {
-        if(fluidStack == null) {
+        if (fluidStack == null) {
             return null;
         }
         Fluid fluid = fluidStack.getFluid();
-        if(fluid == null) {
+        if (fluid == null) {
             return null;
         }
         Map<Fluid, int[]> cache = flowing ? flowingRenderCache : stillRenderCache;
         int[] diplayLists = cache.get(fluid);
-        if(diplayLists != null) {
+        if (diplayLists != null) {
             return diplayLists;
         }
 
         diplayLists = new int[DISPLAY_STAGES];
 
-        if(fluid.getBlockID() > 0) {
+        if (fluid.getBlockID() > 0) {
             liquidBlock.baseBlock = Block.blocksList[fluid.getBlockID()];
             liquidBlock.texture = getFluidTexture(fluidStack, flowing);
-        }else {
+        } else {
             liquidBlock.baseBlock = Block.waterStill;
             liquidBlock.texture = getFluidTexture(fluidStack, flowing);
         }
@@ -95,7 +94,7 @@ public class FluidRender {
         GL11.glDisable(GL11.GL_BLEND);
         GL11.glDisable(GL11.GL_CULL_FACE);
 
-        for(int s = 0; s < DISPLAY_STAGES; ++s) {
+        for (int s = 0; s < DISPLAY_STAGES; ++s) {
             diplayLists[s] = GLAllocation.generateDisplayLists(1);
             GL11.glNewList(diplayLists[s], 4864 /* GL_COMPILE */);
